% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cumerr_plot}
\alias{cumerr_plot}
\title{Cumulative relative error for reaches}
\usage{
cumerr_plot(valdf, reach_ids = "all", variables = c("wse", "width",
  "area_total"), desc = FALSE, plot = TRUE)
}
\arguments{
\item{valdf}{As returned by \code{rt_valdata()}}

\item{reach_ids}{Vector of \code{reach_id}s to include, or "all"}

\item{variables}{Vector of variables to include, or "all"}

\item{desc}{If TRUE, sort x-axis by descending node-level relative error}

\item{plot}{if FALSE, return the data used to plot, but do not generate the plot.}
}
\description{
Plots cumulative relative error against node index, sorted by
 node-level relative error magnitude.
}
