% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{looerr_plot}
\alias{looerr_plot}
\title{Leave-one-out relative error for reaches}
\usage{
looerr_plot(valdf, reach_ids = "all", variables = c("wse", "width",
  "area_total"), sort = c("id", "relerr"), plot = TRUE)
}
\arguments{
\item{valdf}{As returned by \code{rt_valdata()}}

\item{reach_ids}{Vector of \code{reach_id}s to include, or "all"}

\item{variables}{Vector of variables to include, or "all"}

\item{sort}{How to sort the x-axis? Choices are: \code{id} for node_id,
or \code{relerr} for relative error.}

\item{plot}{if FALSE, return the data used to plot, but do not generate the plot.}
}
\description{
Plots total relative error against nodes, excluding that node from
 the calculation of reach-level relative error.
}
