% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{mismatch_nodes}
\alias{mismatch_nodes}
\title{Flag nodes that have either total or potential partial mismatch between gdem and pixc.}
\usage{
mismatch_nodes(obs, truth)
}
\arguments{
\item{obs}{data.frame with observed node-level values}

\item{truth}{data.frame with node-level truth values}
}
\description{
Flag nodes that have either total or potential partial mismatch between gdem and pixc.
}
