% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{nodearea_plot}
\alias{nodearea_plot}
\title{Plot the area computation for a node}
\usage{
nodearea_plot(pixc_joined, nodes, node_truth = NULL, plot = TRUE)
}
\arguments{
\item{pixc_joined}{A joined pixc data frame, as returned by \code{join_pixc()}}

\item{nodes}{Vector giving indices of nodes to plot}

\item{node_truth}{Optional truth data, as returned by \code{rt_read()}}

\item{plot}{if FALSE, return the plot data but don't construct the ggplot}
}
\description{
Plot the area computation for a node
}
