% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rt_cumulplot}
\alias{rt_cumulplot}
\title{Plot error accmulation from nodes to reaches}
\usage{
rt_cumulplot(nodevaldf, var = c("wse", "width", "area_total"),
  scale = c("none", "truth", "unc"), plot = TRUE)
}
\arguments{
\item{nodevaldf}{as returned by \code{rt_valdata(group = "node")}}

\item{var}{variable to accumulate}

\item{scale}{what to scale errors by--"none" for raw errors}

\item{plot}{if FALSE return the plot data only}
}
\description{
Plot error accmulation from nodes to reaches
}
