% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rt_val_hist}
\alias{rt_val_hist}
\title{Plot a histogram of errors, possibly scaled by uncertainty estimates}
\usage{
rt_val_hist(valdata, center = FALSE, scale = FALSE, curve = center &&
  scale, vars = "all", plot = TRUE, ...)
}
\arguments{
\item{valdata}{data.frame as returned by code{rt_valdata()}}

\item{center}{Subtract the mean (bias-correct) the errors?}

\item{scale}{Scale the errors by 1-sigma uncertainty estimates?}

\item{curve}{Overlay a standard normal curve? This is the default if
both \code{center} and \code{scale} are \code{TRUE}.}

\item{vars}{Which variables to plot? Defaults to "all"}

\item{plot}{if FALSE, return the data used to plot but do not create the plot}

\item{...}{Passed to \code{geom_histogram()}}
}
\description{
Currently only implemented for node-level data
}
