% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rt_val_scatter}
\alias{rt_val_scatter}
\title{Validation scatterplot}
\usage{
rt_val_scatter(valdata, variables = c("wse", "width", "slope"),
  xvar = "id", yvar = c("value", "err", "relerr"), ci1 = 0.6827,
  ci2 = 0.95, plot = TRUE)
}
\arguments{
\item{valdata}{As returned by \code{rt_valdata()}}

\item{variables}{Variables to plot}

\item{xvar, yvar}{x- and y-axis variables}

\item{ci1, ci2}{Confidence-interval levels for 2 ribbon geoms.
Either or can be set to FALSE to disable plotting.}

\item{plot}{if FALSE, just return the data used for plotting}
}
\description{
Validation scatterplot
}
