% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{rt_valdata}
\alias{rt_valdata}
\title{Get a validation dataset from a set of RiverObs runs}
\usage{
rt_valdata(dir, group = c("nodes", "reaches"), rtname = "rt.nc",
  gdname = "rt_gdem.nc", keep_na_vars = FALSE,
  time_round_digits = -2, flag_out_nodes = TRUE)
}
\arguments{
\item{dir}{directory containing riverobs output including gdem truth}

\item{group}{Which group to get data from: "nodes" or "reaches}

\item{rtname}{Name of rivertile file to read}

\item{gdname}{Name of gdem-truth rivertile file}

\item{keep_na_vars}{Keep variables that only contain missing values?}

\item{time_round_digits}{how many digits to round time (secondes) to
consider equal between gdem pixc and "real" pixc}

\item{flag_out_nodes}{Automatically flag and remove nodes with ambiguous truth?}
}
\description{
Get a validation dataset from a set of RiverObs runs
}
