% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{rt_valdata_df}
\alias{rt_valdata_df}
\title{Create validation data from two node-level rivertile data.frames.}
\usage{
rt_valdata_df(obs, truth, time_round_digits = -2)
}
\arguments{
\item{obs}{data.frame with observed node-level values}

\item{truth}{data.frame with node-level truth values}

\item{time_round_digits}{how many digits to round time (secondes) to
consider equal between gdem pixc and "real" pixc}
}
\description{
Used internally by \code{rt_valdata()}
}
