% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{val_coverage}
\alias{val_coverage}
\title{Table of coverage for various (normal) confidence intervals.}
\usage{
val_coverage(valdata, ci = c(68, 90, 95, 99), debias = FALSE)
}
\arguments{
\item{valdata}{as returned by \code{rt_valdata()}}

\item{ci}{Which confidence levels to validate?}

\item{debias}{Remove bias error component before validating?}
}
\description{
Table of coverage for various (normal) confidence intervals.
}
