% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{val_map_node}
\alias{val_map_node}
\title{Map a given number of nodes' pixcvec locations--gdem versus rivertile}
\usage{
val_map_node(dir, nodes = badnodes(rt_valdata(dir)), pcv1 = "pcv.nc",
  pcv2 = "pcv_gdem.nc", pixc1 = "pixel_cloud.nc",
  pixc2 = "fake_pixc.nc", maxpixels = 1000)
}
\arguments{
\item{dir}{A directory containing rivertile output.}

\item{nodes}{A vector of node_id's, defaults to worst nodes using \code{badnodes()}}

\item{pcv1, pcv2}{Name of pixcvec netcdfs}

\item{pixc1, pixc2}{Name of pixel cloud netcdfs}

\item{maxpixels}{Maximum number of pixels to plot, as a random sample if necessary.}
}
\description{
Map a given number of nodes' pixcvec locations--gdem versus rivertile
}
